+++ www-admin/js/components/AdminApp.lit.js
@@ -1,7 +1,7 @@
 import { LitElement, html, css } from '/static/js/vendor/lit.js';
 import './admin/GlobalSettings.lit.js';
 import './admin/AzureDevOps.lit.js';
-import './admin/Events.lit.js';
+import './admin/DataSources.lit.js';
 
 export class AdminApp extends LitElement {
   static properties = {
@@ -88,7 +88,7 @@ export class AdminApp extends LitElement {
     this.sections = [
       'Server',
       'Azure DevOps',
-      'Events',
+      'Data Sources',
       'Users',
       'Global Project Settings',
       'Projects',
@@ -137,8 +137,8 @@ export class AdminApp extends LitElement {
         return html`<admin-system></admin-system>`;
       case 'Azure DevOps':
         return html`<admin-ado></admin-ado>`;
-      case 'Events':
-        return html`<admin-events></admin-events>`;
+      case 'Data Sources':
+        return html`<admin-data-sources></admin-data-sources>`;
       case 'Users':
         return html`<admin-users></admin-users>`;
       case 'Global Project Settings':